потёр лишнее вынес TopBar в отдельную папку
