+++ src/options/TopBar/TopBar.css
@@ -1,0 +1,67 @@
+.TopBar__headerGroup {
+   display: flex;
+   align-items: baseline;
+}
+
+.TopBar__header,
+.TopBar__subHeader {
+   letter-spacing: .5px;
+   text-transform: uppercase;
+   font-weight: bold;
+   padding-right: 8px;
+}
+
+.TopBar__header {
+   color: #1e90ff;
+   font-size: 28px;
+}
+
+.TopBar__subHeader {
+   font-size: 18px;
+   color: #9db9d4;
+}
+
+.TopBar__exitGroup {
+   display: flex;
+   align-items: center;
+}
+
+.TopBar__exitGroup:hover {
+   text-decoration: underline;
+   cursor: pointer;
+}
+
+.TopBar__logout {
+   padding: 0 10px;
+   color: #425a70;
+   text-transform: uppercase;
+   line-height: 16px;
+   font-size: 12px;
+   width: 32px;
+   padding-top: 4px;
+}
+
+.TopBar__logoBlock {
+   width: 30px;
+   padding: 0 10px 0 0;
+}
+
+.TopBar__logo {
+   width: 30px;
+   top: 12px;
+   position: absolute;
+}
+
+.TopBar {
+   display: flex;
+   justify-content: space-between;
+   height: 56px;
+   align-items: center;
+   padding-left: 16px;
+   padding-right: 16px;
+   position: sticky;
+   top: 0;
+   z-index: 9;
+   background: #fff;
+   box-shadow: 0 0 1px rgba(67,90,111,.47), 0 2px 4px -2px rgba(67,90,111,.3);
+}
