+++ src/options/User/User.css
@@ -1,71 +1,3 @@
-.User-TopBar__headerGroup {
-    display: flex;
-    align-items: baseline;
-}
-
-.User-TopBar__header,
-.User-TopBar__subHeader {
-    letter-spacing: .5px;
-    text-transform: uppercase;
-    font-weight: bold;
-    padding-right: 8px;
-}
-
-.User-TopBar__header {
-    color: #1e90ff;
-    font-size: 28px;
-}
-
-.User-TopBar__subHeader {
-    font-size: 18px;
-    color: #9db9d4;
-}
-
-.User-TopBar__exitGroup {
-    display: flex;
-    align-items: center;
-}
-
-.User-TopBar__exitGroup:hover {
-    text-decoration: underline;
-    cursor: pointer;
-}
-
-.User-TopBar__logout {
-    padding: 0 10px;
-    color: #425a70;
-    text-transform: uppercase;
-    line-height: 16px;
-    font-size: 12px;
-    width: 32px;
-    padding-top: 4px;
-}
-
-.User-TopBar__logoBlock {
-    width: 30px;
-    padding: 0 10px 0 0;
-}
-
-.User-TopBar__logo {
-    width: 30px;
-    top: 12px;
-    position: absolute;
-}
-
-.User-TopBar {
-    display: flex;
-    justify-content: space-between;
-    height: 56px;
-    align-items: center;
-    padding-left: 16px;
-    padding-right: 16px;
-    position: sticky;
-    top: 0;
-    z-index: 9;
-    background: #fff;
-    box-shadow: 0 0 1px rgba(67,90,111,.47), 0 2px 4px -2px rgba(67,90,111,.3);
-}
-
 .User-LearningPage__height-90vh {
     min-height: 90vh;
 }
